Move get_json to the place where it is actually used
